Reworked WSPBus plugins; most now have a 'subscribe' method which attaches them to the bus in a separate step from init. This allows frameworks to create canonical plugin instances but delay or skip their subscription.
Some plugins' method names changed (from e.g. {{{__call__}}} to {{{start}}}) to take advantage of a new SimplePlugin base class. The daemonize func is now a Daemonizer class for this reason.

Also removed the SubscribedObject class; it was too much magic. The few consumers of it in CherryPy revert to calling canonical plugin instances instead.

git-svn-id: http://svn.cherrypy.org/trunk@1752 e1d34091-3ce9-0310-8e96-997e60db3bd5